more progress about 50-69% done
first enemy spawns and uses a "chase" movement system to hone in on the enemy using coordinates in the x,y plane. Updated the Tear class to use the correct class and added a cooldown so only a few tears can be shot at any time (so the form doesn't lag). Also, max enemies is now set to 4.