Menu system complete (basically)
Almost fully finished the menu system (just need the game over menu which will have quit, new game, and scores options or can take you to the main menu. This will be done later when I revisit combat and the rest of the game mechanics. but for now main menu works :3